tests/wpt/web-platform-tests/webmessaging/without-ports/009.html

<!doctype html>
<title>zero arguments</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
test(function() {
  assert_throws(new TypeError(), function() {
    postMessage();
  });
});
</script>
